requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pypdf==3.17.4
python-docx==1.1.0
edge-tts==6.1.9
langdetect==1.0.9
sumy==0.11.0
nltk==3.8.1
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.26.0